chore(core, collect, manage): refactor core modules structure, modularize services, update dependency configurations, enhance CLI initialization, and add `manage` crate
