+++ crates/collect/Cargo.toml
@@ -17,7 +17,6 @@ path = "src/main.rs"
 [dependencies]
 ammonia = { version = "^4.1.2" }
 async-trait.workspace = true
-dotenvy.workspace = true
 futures.workspace = true
 htmd = { version = "^0.3.2" }
 rss = { version = "^2.0.12" }
@@ -25,7 +24,6 @@ serde.workspace = true
 tokio.workspace = true
 toml.workspace = true
 tracing.workspace = true
-tracing-subscriber.workspace = true
 uninews_core = { path = "../core" }
 url.workspace = true
 uuid.workspace = true
