+++ crates/collect/src/cli/runners.rs
@@ -1,6 +1,6 @@
 use crate::state::AppState;
 use futures::future::join_all;
-use uninews_core::source::Source;
+use uninews_core::models::Source;
 
 pub async fn run_collectors(app_state: AppState) {
     join_all(app_state.sources().map(Source::watch_updates)).await;
