+++ crates/core/Cargo.toml
@@ -16,9 +16,11 @@ path = "src/lib.rs"
 
 [dependencies]
 async-trait.workspace = true
+dotenvy.workspace = true
 futures.workspace = true
 reqwest.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+tracing-subscriber.workspace = true
 url.workspace = true
 uuid.workspace = true
