+++ crates/core/src/fs.rs
