--- crates/core/src/http.rs
@@ -1,2 +1,0 @@
-pub mod live;
-pub mod service;
