+++ crates/core/src/services.rs
@@ -1,0 +1,3 @@
+pub mod http;
+pub mod news;
+pub mod storage;
