--- crates/core/src/storage.rs
@@ -1,2 +1,0 @@
-pub mod live;
-pub mod service;
