--- crates/core/src/utils.rs
@@ -1,2 +1,0 @@
-pub mod fs;
-pub mod uuid;
