+++ crates/core/src/uuid.rs
