+++ crates/manage/Cargo.toml
@@ -15,3 +15,6 @@ name = "manage"
 path = "src/main.rs"
 
 [dependencies]
+tokio.workspace = true
+tracing.workspace = true
+uninews_core = { path = "../core" }
