+++ crates/manage/src/main.rs
@@ -1,3 +1,6 @@
-fn main() {
-    println!("Hello, world!");
+use uninews_manage::cli::run;
+
+#[tokio::main(flavor = "multi_thread")]
+async fn main() {
+    run().await;
 }
